RL-A2C/requirements.txt
gymnasium==0.29.1
stable-baselines3==2.1.0
paho-mqtt==1.6.1
numpy==1.26.4
pyyaml==6.0.1
matplotlib==3.8.2
tqdm==4.66.1
setuptools>=65.5.0